Разбил код файла form (выполняющий добавление и редактирования фотографий) на модули
